Sound for walking is now in the animation
