+++ Assets/Scripts/Behaviors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: beea84770b9b9e24a849136e61dc2696
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
