+++ Assets/Scripts/Behaviors/LandAnimationStateBehaviour.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+
+namespace Behaviors
+{
+    public class LandAnimationStateBehaviour : StateMachineBehaviour
+    {
+        private GameObject _audioController;
+
+        private void Awake()
+        {
+            _audioController = GameObject.Find("AudioController");
+        }
+        public override void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+        {
+            _audioController.GetComponent<SFX>().PlayLanding();
+        }
+    }
+}
