+++ Assets/Scripts/Behaviors/WalkAnimationCooldown.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3eecd14dd4ac39247b9fce05af39bffb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
